Implemented star logic & design. Refactored code
