+++ .gitignore
@@ -2,6 +2,7 @@
 #
 # Get latest from https://github.com/github/gitignore/blob/main/Unity.gitignore
 #
+Assets/TeKniKo_Free
 /Assets/DavePixel
 /Assets/Game Sound Solutions - 8 bits Elements
 /[Ll]ibrary/
