+++ Assets/Audio/coin_11.wav.meta
@@ -1,0 +1,30 @@
+fileFormatVersion: 2
+guid: 2e1c0edfc97324f3cbdc4ed5f343753b
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 8
+  defaultSettings:
+    serializedVersion: 2
+    loadType: 1
+    sampleRateSetting: 0
+    sampleRateOverride: 0
+    compressionFormat: 0
+    quality: 1
+    conversionMode: 0
+    preloadAudioData: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+AssetOrigin:
+  serializedVersion: 1
+  productId: 16848
+  packageName: 8 Bits Elements
+  packageVersion: 1.3.1
+  assetPath: Assets/Game Sound Solutions - 8 bits Elements/coin/coin_11.wav
+  uploadId: 46669
