+++ Assets/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fa32ab2c60b37ea468781803e4d1a2c3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
