+++ Assets/Materials/Gold.mat.meta
@@ -1,0 +1,15 @@
+fileFormatVersion: 2
+guid: 3712460f643bb3147889765d028e1c84
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+AssetOrigin:
+  serializedVersion: 1
+  productId: 207743
+  packageName: Pirate Coin
+  packageVersion: 1.0
+  assetPath: Assets/DavePixel/PirateCoin/Materials/Gold.mat
+  uploadId: 467011
