+++ Assets/Scripts/Collectible.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 95d10e464c4212142bb0cdc243de5d28
