--- Assets/Scripts/coinTest.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b99582930a1ea1b43a7e79cdf5040aad
