+++ Assets/TeKniKo_Free.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d9896cf4ad9be3843a275d9865a6bbd2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
