+++ Assets/Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6062ac13471971b44bd64b889c26f21c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
